+++ packages/router/src/handler.ts
@@ -13,18 +13,18 @@ import {
   NxtpError,
   Values,
   NxtpErrorJson,
-  recoverAuctionBid,
+  getUuid,
+  RequestContext,
   decodeAuctionBid,
+  recoverAuctionBid,
 } from "@connext/nxtp-utils";
 import { BigNumber, utils, Wallet } from "ethers";
-import hyperid from "hyperid";
 import { combine, err, errAsync, ok, okAsync, Result, ResultAsync } from "neverthrow";
 import { BaseLogger } from "pino";
 
 import { getConfig } from "./config";
 import { TransactionManager } from "./contract";
 import { Subgraph } from "./subgraph";
-
 /**
  * Handler.ts
 
@@ -38,8 +38,6 @@ import { Subgraph } from "./subgraph";
   4. Dispatch a new message or tx to chain
   5. Update metrics
  */
-
-const hId = hyperid();
 
 export const EXPIRY_DECREMENT = 3600 * 24;
 export const SWAP_RATE = "0.9995"; // 0.05% fee
@@ -66,6 +64,7 @@ export class HandlerError extends NxtpError {
     TxServiceError: "Error from Transaction Service",
     TxAlreadyPrepared: "Transaction already prepared",
     AuctionValidationError: "Error validating auction params",
+    ConfigError: "Config error",
     EncodeError: "Error encoding",
     PrepareValidationError: "Error validating sender prepare",
   };
@@ -76,6 +75,8 @@ export class HandlerError extends NxtpError {
       auctionError?: NxtpErrorJson;
       messagingError?: NxtpErrorJson;
       txServiceError?: NxtpErrorJson;
+      configError?: string;
+      requestContext: RequestContext;
       encodingError?: NxtpErrorJson;
       prepareError?: string;
       methodId: string;
@@ -86,6 +87,17 @@ export class HandlerError extends NxtpError {
     super(message, context, HandlerError.type);
   }
 }
+
+/**
+ * This is only here to make it easier for sinon mocks to happen in the tests. Otherwise, this is a very dumb thing.
+ *
+ * @param bid - Bid information that should've been signed
+ * @param signature - Signature to recover signer of
+ * @returns Recovered signer
+ */
+export const recoverAuctionSigner = (bid: AuctionBid, signature: string): string => {
+  return recoverAuctionBid(bid, signature);
+};
 
 /**
  * Returns the amount * SWAP_RATE to deduct fees when going from sending -> recieving chain to incentivize routing.
@@ -128,6 +140,8 @@ export const getBidExpiry = () => Math.floor(Date.now() / 1000) + 60 * 5;
  * Contains all callback logic for events and messages received by the router
  */
 export class Handler {
+  private receiverPreparing: Map<string, boolean> = new Map();
+  private senderFulfilling: Map<string, boolean> = new Map();
   constructor(
     private readonly messagingService: RouterNxtpNatsMessagingService,
     private readonly _subgraph: Subgraph,
@@ -159,10 +173,10 @@ export class Handler {
    * @param inbox - The inbox to publish the auction response to (where the initiator of the auction will be listening for bids)
    * @returns An empty promise
    */
-  public async handleNewAuction(data: AuctionPayload, inbox: string): Promise<void> {
+  public async handleNewAuction(data: AuctionPayload, inbox: string, requestContext: RequestContext): Promise<void> {
     const method = this.handleNewAuction.name;
-    const methodId = hId();
-    this.logger.info({ method, methodId, data, inbox }, "Method start");
+    const methodId = getUuid();
+    this.logger.info({ method, methodId, requestContext, data, inbox }, "Method start");
 
     const {
       user,
@@ -199,6 +213,7 @@ export class Handler {
               calling: "",
               methodId,
               method,
+              requestContext,
               auctionError: {
                 message: "Not enough availble liquidity for auction",
                 type: "Validation",
@@ -228,6 +243,7 @@ export class Handler {
               calling: "",
               methodId,
               method,
+              requestContext,
               auctionError: {
                 message: "Providers not available for both chains",
                 type: "Validation",
@@ -256,6 +272,7 @@ export class Handler {
               calling: "",
               methodId,
               method,
+              requestContext,
               auctionError: {
                 message: "Allowed swap not part of config",
                 type: "Validation",
@@ -278,6 +295,7 @@ export class Handler {
                 calling: "txService.getBalance => sending",
                 method,
                 methodId,
+                requestContext,
                 txServiceError: jsonifyError(err as NxtpError),
               }),
           ),
@@ -288,6 +306,7 @@ export class Handler {
                 calling: "txService.getBalance => receiving",
                 method,
                 methodId,
+                requestContext,
                 txServiceError: jsonifyError(err as NxtpError),
               }),
           ),
@@ -301,6 +320,7 @@ export class Handler {
               calling: "",
               methodId,
               method,
+              requestContext,
               auctionError: {
                 message: "Not enough gas",
                 type: "Validation",
@@ -315,7 +335,7 @@ export class Handler {
         return okAsync({ senderBalance, receiverBalance });
       })
       .andThen(() => {
-        this.logger.info({ method, methodId }, "Auction validation complete, generating bid");
+        this.logger.info({ method, methodId, requestContext }, "Auction validation complete, generating bid");
         // (TODO in what other scenarios would auction fail here? We should make sure
         // that router does not bid unless it is *sure* it's doing ok)
         // If you can support the transfer:
@@ -345,7 +365,7 @@ export class Handler {
           receivingChainTxManagerAddress: receivingConfig.transactionManagerAddress,
           bidExpiry: getBidExpiry(),
         };
-        this.logger.info({ methodId, method, bid }, "Generated bid");
+        this.logger.info({ methodId, method, requestContext, bid }, "Generated bid");
 
         return ResultAsync.fromPromise(
           signAuctionBid(bid, this.signer),
@@ -353,19 +373,21 @@ export class Handler {
             new HandlerError(HandlerError.reasons.AuctionValidationError, {
               calling: "signAuctionBid",
               methodId,
+              requestContext,
               method,
               auctionError: jsonifyError(err as Error),
             }),
         );
       })
       .andThen((bidSignature) => {
-        this.logger.info({ methodId, method, bidSignature }, "Signed bid");
+        this.logger.info({ methodId, method, requestContext, bidSignature }, "Signed bid");
         return ResultAsync.fromPromise(
           this.messagingService.publishAuctionResponse(inbox, { bid, bidSignature: dryRun ? undefined : bidSignature }),
           (err) =>
             new HandlerError(HandlerError.reasons.MessagingError, {
               calling: "messagingService.publishAuctionResponse",
               methodId,
+              requestContext,
               method,
               messagingError: jsonifyError(err as Error),
             }),
@@ -374,9 +396,9 @@ export class Handler {
 
     if (result.isOk()) {
       // Last, update metrics
-      this.logger.info({ method, methodId, result: result.value }, "Method complete");
+      this.logger.info({ method, methodId, requestContext, result: result.value }, "Method complete");
     } else {
-      this.logger.error({ method, methodId, transactionId, err: result.error });
+      this.logger.error({ method, methodId, requestContext, transactionId, err: result.error }, "Error in method");
     }
   }
 
@@ -398,15 +420,42 @@ export class Handler {
    * The transaction should be broadcast to an entire network of relayers who will race to claim the fee. If it is only broadcast to a *single* relayer, then they could collude with the transaction's router to cancel the receiver-side payment and fulfill the sender-side payment (since the signature has been revealed)
    *
    */
-  public async handleMetaTxRequest(data: MetaTxPayload<any>, inbox: string): Promise<void> {
+  public async handleMetaTxRequest(
+    data: MetaTxPayload<any>,
+    inbox: string,
+    requestContext: RequestContext,
+  ): Promise<void> {
     // First log
     const method = this.handleMetaTxRequest.name;
-    const methodId = hId();
-    this.logger.info({ method, methodId, data }, "Method start");
+    const methodId = getUuid();
+    this.logger.info({ method, methodId, requestContext, data }, "Method start");
 
     const { chainId } = data;
+    const config = getConfig();
+    const chainConfig = config.chainConfig[chainId];
+    if (!chainConfig) {
+      const err = new HandlerError(HandlerError.reasons.ConfigError, {
+        requestContext,
+        calling: "getConfig",
+        methodId,
+        method,
+        configError: `No chainConfig for ${chainId}`,
+      });
+      this.logger.error({ method, methodId, requestContext, err: err.toJson() }, "Error in config");
+    }
 
     if (data.type === "Fulfill") {
+      if (utils.getAddress(data.to) !== utils.getAddress(chainConfig.transactionManagerAddress)) {
+        const err = new HandlerError(HandlerError.reasons.ConfigError, {
+          requestContext,
+          calling: "chainConfig.transactionManagerAddress",
+          methodId,
+          method,
+          configError: `Provided transactionManagerAddress does not map to our configured transactionManagerAddress`,
+        });
+        this.logger.error({ method, methodId, requestContext, err: err.toJson() }, "Error in config");
+      }
+
       const fulfillData: MetaTxFulfillPayload = data.data;
       // Validate that metatx request matches with known data about fulfill
       // Is this needed? Can we just submit to chain without validating?
@@ -422,32 +471,41 @@ export class Handler {
 
       // TODO: make sure fee is something we want to accept
 
-      this.logger.info({ method, methodId, chainId, data }, "Submitting tx");
+      this.logger.info({ method, methodId, requestContext, chainId, data }, "Submitting tx");
       const res = await this.txManager
-        .fulfill(chainId, {
-          txData: fulfillData.txData,
-          relayerFee: fulfillData.relayerFee,
-          signature: fulfillData.signature,
-          callData: fulfillData.callData,
-        })
+        .fulfill(
+          chainId,
+          {
+            txData: fulfillData.txData,
+            relayerFee: fulfillData.relayerFee,
+            signature: fulfillData.signature,
+            callData: fulfillData.callData,
+          },
+          requestContext,
+        )
         .andThen((tx) => {
-          this.logger.info({ method, methodId, transactionHash: tx.transactionHash }, "Relayed transaction");
+          this.logger.info(
+            { method, methodId, requestContext, transactionHash: tx.transactionHash },
+            "Relayed transaction",
+          );
           return ResultAsync.fromPromise(
             this.messagingService.publishMetaTxResponse(inbox, { chainId, transactionHash: tx.transactionHash }),
             (err) =>
               new HandlerError(HandlerError.reasons.MessagingError, {
                 method,
                 methodId,
+                requestContext,
                 calling: "messagingService.publishMetaTxResponse",
                 messagingError: jsonifyError(err as Error),
               }),
           );
         })
         .orElse((err) => {
-          this.logger.info({ method, methodId, transactionHash: "" }, "Error relaying transaction");
+          this.logger.error({ method, methodId, requestContext }, "Error relaying transaction");
           const _err = new HandlerError(HandlerError.reasons.TxServiceError, {
             method,
             methodId,
+            requestContext,
             calling: "txManager.fulfill",
             txServiceError: jsonifyError(err),
           });
@@ -456,6 +514,7 @@ export class Handler {
             (err) =>
               new HandlerError(HandlerError.reasons.MessagingError, {
                 method,
+                requestContext,
                 methodId,
                 calling: "messagingService.publishMetaTxResponse",
                 messagingError: jsonifyError(err as Error),
@@ -464,9 +523,9 @@ export class Handler {
         });
 
       if (res.isOk()) {
-        this.logger.info({ method, methodId }, "Method complete");
+        this.logger.info({ method, methodId, requestContext }, "Method complete");
       } else {
-        this.logger.error({ method, methodId, err: jsonifyError(res.error) }, "Error occurred");
+        this.logger.error({ method, methodId, requestContext, err: jsonifyError(res.error) }, "Error occurred");
       }
     }
   }
@@ -486,12 +545,23 @@ export class Handler {
    * @remarks
    * Should cancel the transsaction on the sending chain if it fails to prepare on the receiving chain. Will also prepare with a lower amount and lower expiry then what was prepared with on the receiving chain.
    */
-  public async handleSenderPrepare(inboundData: TransactionPreparedEvent): Promise<void> {
+  public async handleSenderPrepare(
+    inboundData: TransactionPreparedEvent,
+    requestContext: RequestContext,
+  ): Promise<void> {
     const method = "handleSenderPrepare";
-    const methodId = hId();
-    this.logger.info({ method, methodId, inboundData }, "Method start");
+    const methodId = getUuid();
+    this.logger.info({ method, methodId, inboundData, requestContext }, "Method start");
 
     const { txData, bidSignature, encodedBid, encryptedCallData, amount } = inboundData;
+
+    if (this.receiverPreparing.get(txData.transactionId)) {
+      this.logger.info({ methodId, method, requestContext, transactionId: txData.transactionId }, "Already fulfilling");
+      return;
+    }
+
+    // TODO: what if theres never a fulfill, where does receiver cancellation
+    // get handled? sender + receiver cancellation?
 
     // Validate the prepare data
     // TODO what needs to be validated here? Is this necessary? Assumption
@@ -506,18 +576,21 @@ export class Handler {
           method,
           methodId,
           calling: "decodeAuctionBid",
+          requestContext,
           encodingError: jsonifyError(err as Error),
         }),
     )(encodedBid)
       .andThen((_bid) => {
         bid = _bid;
+        this.logger.info({ method, methodId, requestContext, bid }, "Decoded bid from event");
         return Result.fromThrowable(
-          recoverAuctionBid,
+          recoverAuctionSigner,
           (err) =>
             new HandlerError(HandlerError.reasons.EncodeError, {
               method,
               methodId,
-              calling: "recoverAuctionBid",
+              calling: "recoverAuctionSigner",
+              requestContext,
               encodingError: jsonifyError(err as Error),
             }),
         )(bid, bidSignature);
@@ -529,6 +602,7 @@ export class Handler {
               method,
               methodId,
               calling: "",
+              requestContext,
               prepareError: "recovered !== this.signer.address",
             }),
           );
@@ -537,12 +611,13 @@ export class Handler {
       })
       .andThen(() => {
         // TODO: anything else? seems unnecessary to validate everything
-        if (bid.amount !== amount || bid.transactionId !== txData.transactionId) {
+        if (!BigNumber.from(bid.amount).eq(txData.amount) || bid.transactionId !== txData.transactionId) {
           return err(
             new HandlerError(HandlerError.reasons.PrepareValidationError, {
               method,
               methodId,
               calling: "",
+              requestContext,
               prepareError: "Bid params not equal to tx data",
             }),
           );
@@ -551,13 +626,19 @@ export class Handler {
       });
 
     if (validationRes.isOk()) {
-      //
+      this.logger.info({ method, methodId, requestContext }, "Validated input");
     } else {
       this.logger.error(
-        { method, methodId, transactionId: txData.transactionId, err: jsonifyError(validationRes.error as Error) },
+        {
+          method,
+          methodId,
+          requestContext,
+          transactionId: txData.transactionId,
+          err: jsonifyError(validationRes.error as Error),
+        },
         "Error during validation",
       );
-      // TODO: maybe cancel here
+      return;
     }
 
     // Next, prepare the outbound data
@@ -581,18 +662,30 @@ export class Handler {
     // - AssetId
     // encode the data for contract call
     // Send to txService
+    this.receiverPreparing.set(txData.transactionId, true);
+    this.logger.info(
+      { method, methodId, requestContext, transactionId: txData.transactionId },
+      "Sending receiver prepare tx",
+    );
+    const res = await this.txManager.prepare(
+      txData.receivingChainId,
+      {
+        txData,
+        amount: mutateAmount(txData.amount),
+        expiry: mutateExpiry(txData.expiry),
+        bidSignature,
+        encodedBid,
+        encryptedCallData,
+      },
+      requestContext,
+    );
     this.logger.info({ method, methodId, transactionId: txData.transactionId }, "Sending receiver prepare tx");
-    const res = await this.txManager.prepare(txData.receivingChainId, {
-      txData,
-      amount: mutateAmount(amount),
-      expiry: mutateExpiry(txData.expiry),
-      bidSignature,
-      encodedBid,
-      encryptedCallData,
-    });
 
     if (res.isOk()) {
-      this.logger.info({ method, methodId, transactionHash: res.value.transactionHash }, "Prepared transaction");
+      this.logger.info(
+        { method, methodId, requestContext, transactionHash: res.value.transactionHash },
+        "Prepared transaction",
+      );
       // If success, update metrics
     } else {
       if (res.error.message.includes("#P:015")) {
@@ -600,6 +693,7 @@ export class Handler {
           {
             method,
             methodId,
+            requestContext,
             transactionId: txData.transactionId,
           },
           "Tx Failed because it was already prepared, do not cancel",
@@ -610,6 +704,7 @@ export class Handler {
           {
             method,
             methodId,
+            requestContext,
             transactionId: txData.transactionId,
             prepareError: jsonifyError(res.error),
           },
@@ -619,6 +714,7 @@ export class Handler {
           {
             method,
             methodId,
+            requestContext,
           },
           "Do not cancel ATM, figure out why we are in this case first",
         );
@@ -637,6 +733,7 @@ export class Handler {
         // }
       }
     }
+    this.receiverPreparing.delete(txData.transactionId);
   }
 
   // HandleReceiverFulfill
@@ -664,28 +761,51 @@ export class Handler {
   public async handleReceiverFulfill(
     senderEvent: TransactionPreparedEvent,
     receiverEvent: TransactionFulfilledEvent,
+    requestContext: RequestContext,
   ): Promise<void> {
     const method = this.handleReceiverFulfill.name;
-    const methodId = hId();
-    this.logger.info({ method, methodId, senderEvent, receiverEvent }, "Method start");
+    const methodId = getUuid();
+    this.logger.info({ method, methodId, requestContext, senderEvent, receiverEvent }, "Method start");
 
     const { txData, signature, callData, relayerFee } = receiverEvent;
 
-    // Send to tx service
-    this.logger.info({ method, methodId, transactionId: txData.transactionId, signature }, "Sending sender fulfill tx");
+    if (this.senderFulfilling.get(txData.transactionId)) {
+      this.logger.info({ methodId, method, requestContext, transactionId: txData.transactionId }, "Already fulfilling");
+      return;
+    }
 
-    const res = await this.txManager.fulfill(txData.sendingChainId, {
-      txData: senderEvent.txData,
-      signature,
-      relayerFee,
-      callData,
-    });
+    this.senderFulfilling.set(txData.transactionId, true);
+
+    // Send to tx service
+    this.logger.info(
+      { method, methodId, requestContext, transactionId: txData.transactionId, signature },
+      "Sending sender fulfill tx",
+    );
+
+    const res = await this.txManager.fulfill(
+      txData.sendingChainId,
+      {
+        txData: senderEvent.txData,
+        signature,
+        relayerFee,
+        callData,
+      },
+      requestContext,
+    );
 
     if (res.isOk()) {
-      this.logger.info({ method, methodId, transactionHash: res.value.transactionHash }, "Fulfilled transaction");
+      this.logger.info(
+        { method, methodId, requestContext, transactionHash: res.value.transactionHash },
+        "Fulfilled transaction",
+      );
       // If success, update metrics
     } else {
-      this.logger.error({ method, methodId, err: jsonifyError(res.error) }, "Fulfilled transaction");
+      this.logger.error(
+        { method, methodId, requestContext, err: jsonifyError(res.error) },
+        "Error fulfilling transaction",
+      );
     }
+
+    this.senderFulfilling.delete(txData.transactionId);
   }
 }