+++ packages/router/test/utils.ts
@@ -6,6 +6,9 @@ import {
   TransactionData,
   TransactionPreparedEvent,
   TransactionFulfilledEvent,
+  AuctionBid,
+  encodeAuctionBid,
+  mkSig,
 } from "@connext/nxtp-utils";
 import { providers, constants } from "ethers";
 
@@ -53,8 +56,8 @@ export const fakeConfig: NxtpRouterConfig = {
     {
       name: "TEST",
       assets: [
-        { assetId: mkAddress("0xb"), chainId: 1337 },
-        { assetId: mkAddress("0xc"), chainId: 1338 },
+        { assetId: mkAddress("0xc"), chainId: 1337 },
+        { assetId: mkAddress("0xf"), chainId: 1338 },
       ],
     },
   ],
@@ -80,6 +83,26 @@ export const variantDataMock: VariantTransactionData = {
   preparedBlockNumber: 1234,
 };
 
+export const auctionBidMock: AuctionBid = {
+  user: invariantDataMock.user,
+  router: invariantDataMock.router,
+  sendingAssetId: invariantDataMock.sendingAssetId,
+  receivingAssetId: invariantDataMock.receivingAssetId,
+  receivingAddress: invariantDataMock.receivingAddress,
+  sendingChainId: invariantDataMock.sendingChainId,
+  receivingChainId: invariantDataMock.receivingChainId,
+  callTo: invariantDataMock.callTo,
+  callDataHash: invariantDataMock.callDataHash,
+  transactionId: invariantDataMock.transactionId,
+  amount: variantDataMock.amount,
+  sendingChainTxManagerAddress: mkAddress("0x1"),
+  receivingChainTxManagerAddress: mkAddress("0x2"),
+  expiry: variantDataMock.expiry,
+  encryptedCallData: "0x",
+  amountReceived: "120",
+  bidExpiry: 123457,
+};
+
 export const txDataMock: TransactionData = {
   ...invariantDataMock,
   ...variantDataMock,
@@ -90,8 +113,8 @@ export const senderPrepareData: TransactionPreparedEvent = {
   amount: "123",
   caller: mkAddress("0xf"),
   encryptedCallData: "0xabc",
-  encodedBid: "0xdef",
-  bidSignature: "0xbca",
+  encodedBid: encodeAuctionBid(auctionBidMock),
+  bidSignature: mkSig("0xeee"),
 };
 
 export const receiverFulfillDataMock: TransactionFulfilledEvent = {